requirements.txt
fastapi>=0.110,<1.0
pydantic>=2.5,<3.0
sentence-transformers>=2.2,<3.0
faiss-cpu>=1.7.4
numpy>=1.23,<2.0
uvicorn[standard]>=0.23,<1.0
